requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16,<4.0

# Only install pywin32 on Windows
pywin32>=305,<400; platform_system=="Windows"

requests>=2.32,<3.0
httpx>=0.28,<0.30
httpcore>=1.0,<2.0

numpy>=1.24,<2.0
pandas>=2.0,<3.0
scikit-learn>=1.6,<2.0
scipy>=1.10,<2.0
matplotlib>=3.7,<4.0
seaborn>=0.13,<1.0

certifi>=2023.7.22,<2025.0
pytz>=2023.3,<2025.0
python-dateutil>=2.9,<3.0
typing_extensions>=4.12,<5.0
packaging>=24.0,<25.0

Flask>=3.0,<4.0
asgiref>=3.8,<4.0

click>=8.1,<9.0
Jinja2>=3.1,<4.0
itsdangerous>=2.2,<3.0
Werkzeug>=3.1,<4.0
gunicorn>=20.1.0,<21.0.0
whitenoise>=6.0.0

dj-database-url==2.1.0
psycopg2-binary==2.9.9